Remove unused code

+++ launcher3quickstep/src/com/android/launcher3/Utilities.java
@@ -164,7 +164,7 @@ public final class Utilities {
     }
 
     public static int pxFromSp(float size, DisplayMetrics metrics) {
-        return (int) Math.round(TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP,
+        return Math.round(TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP,
                 size, metrics));
     }
